Use pre-transformed vertex coordinates for fragment shader
Instead of dividing fragment coordinates by the rect width and height to normalize coordinates:

Send the vertices to the GPU in a 0-1.0 coordinate space,
Make a copy of the coordinates before transform and transform the input vertices
Send both to the fragment shader, the pretransformed coordinates are used for computing light distance. The other used for drawing the rect.

This saves one divide per fragment.