Various Fixes... this time with files in the right directory
Only create and attach the vertex buffer and attribute once instead of every draw.

Now the main light list is created in UBO format  (ie. as a one dimensional array with padding) instead of being converted every draw.

The light transform is moved to vertex shader. Now we just transform the four vertices the lights are hitting instead of the lights themselves. This removes the need to clone the main light list, transform it in CPU, and re-send the UBO every draw. 

That allows the UBO to be sent only when the background changes.

Different UBOs with the same light list still need to be made for each paintObject but that's a limitation of of OpenGL ES 3.0. We can attach the same program to different contexts.