Just some cleanups
Get rid of a few occasions where we're passing variables to functions that don't need to be passed.

We don't need a light array stored long term in JS. Just make it a UBO when it's first created and that's all that's needed.

Turns out that we don't need getBoundingClientRect() in any of the places it was used so it was removed.

We already weren't using the wallpaper width and height in the shader so I got rid of it. We'll probably use it or it's aspect ratio later on but it won't be changed every draw anyway. It would only need to be changed when the wallpaper changes, just like the light list.