Re-organized everything
Monitor.js is no longer used. Workspace creation and rectangle updating was moved to js.js. Background redrawing was moved to scene.js. The monitor object is no longer a class and may not be needed in the future.

Js.js contains everything in functions which are all called by a main function. This might include additional work in the future to make sure shaders don't try to draw anything before they're actually ready to.

shaders.js now contains the strings for all the shaders as well as a new class that contains all the setup for a WebGL2 context so new ones can be created with less code.