More clean up and fixes
Makes the analyzer worker constant instead of creating a new one every time sendtoAnalyze was called.

sendtoAnalyze has the background listener and analyst response pulled out of it. The listener is now created only once in js.js which removes the need to have the code to kill the listener.

The wallpaper, monitor rect, and theme aren't set until the window loads. This is also where the update loop starts.