Put all image analysis code on seperate thread
Before the creation and shader setup for the off-screen WebGL2 canvas all happened on the main thread, as did this reading pixels back to the CPU. A copy of those pixels were then sent to the worker thread along with wallpaper and canvas width and height. Now the main thread just creates an ImageBitmap out of the wallpaper and just transfers ownership of it to the worker thread. This gets rid of the need to send the image and canvas width and height data because the image dimensions are sent as part of the ImageBitmap object and the canvas is setup on the worker thread.

This clean separation of all analysis code has allowed for the removal of the 'sendtoAnalyze" function and allows for future image analysis code to work with a different thumbnail size without making it difficult to switch back to the median_cut code for testing purposes.