Workspaces canvases replaced with SVGs
The workspaces are no longer canvases that are drawn to with shaders. Instead they use the background is created in a hidden inline SVG and the each workspace is it's own inline SVG that references the contents of the first SVG using <use>.

As a result, the createWorkSpace function, it's associated shader, and any code related to workspaces in other functions were removed and the three functions used to create workspaces and lights have now been combined into one.

The "zoom to fit" behavior of the background just uses the built in preserveAspectRatio feature of SVGs. To prevent the background from blanking during the time when the old background is purged and the new one is loaded, a second <image> element is used and the first one is only removed after the second one as fully loaded. Adding a cross fade should be pretty simple to add from here.