Fixes Behavior on Chrome and Edge
Evidentially V8/Blink-based browsers would hear hundreds of load events when a listener is attached to an <image> and it caused hundreds of web sockets, WebGL contexts, etc to be fired up.

They also don't support CSS nesting if the nest begins with a type selector so they couldn't read any of the CSS variables. I'm working around this by adding a class called "htm" to <html>. It's dumb but whatever.