Moved the sRGB to OkLab conversion to CPU
Only Lightness needed to be computed so it didn't need a matrix transformation. Doing it on the CPU allows us to send the value as a uniform straight to the fragment shader instead of processing it in the vertex shader and sending it as a varying to the fragment shader. There's probably performance benefits to that.

Ideally, we could just set the colors using OkLab in CSS and then animate and pull those values directly. Unfortunately, getComputedStyle only returns sRGB values.