Split up resize function
Previously the resize function was being called any time the wallpaper changed. That needlessly reset the canvas size and viewports of every surface.
The only thing that needs to be computed when changed the wallpaper is the scaling factor between the monitor and wallpaper. That code is now in the aspectRatioChanged function.

The other change is that the resize function only changes the viewport canvas size of the stage since it's the only thing that really changes size. Everything else just gets sent new scale and rect uniforms.

The next thing on my todo list is to prevent the update function from needlessly checking the surface's brightness and sending that every update. That only needs to happen during a transition between light and dark mode.