Move Transform from VS to CPU
This reverts an aspect of the previous commit. There's no longer a function that just sends to the scale to the vertex shader (VS). Now the work of the vertex shader has been moved completely to the CPU. The recalculating the size of surfaces is needed so infrequently that moving it to the VS would have just made it more expensive in the long run. In a real world scenario, the rects for stage, dash, and search bar get defined at login and never change until the user logs out. When it's part of the vertex shader then it it needs to get computed for each surface that is part of a frame update. 

Right now the scale and rects get recomputed when the user switches background's, too and that feels like a reversion from last commit but really the VS was just hiding the fact that it was being done anyway. In fact, it was worse because the VS did for every single update. Resized the window? It recomputed the rects. Changed the background? Resized the rects. Toggle dark mode? That's was the worst one because nothing changes size at all but it has an animation so it does it about 18 times on a 60hz monitor.

Now it gets computed once when resizing or changing the background just like before but nothing gets recomputed when changing to dark or light mode... well, except for the brightness of the surfaces. And I made good on what I said from the last commit. Now it fetches and converts the surface brightness when switching modes. I only did it with an if statement but I did. Hopefully in the future I'll think of something cleaner.