Improve Default Background Code
We place an <image>  with the href set to the default dataURL. setBackground reads from that attribute if it was passed a null background. This gets rid of the branch in analyst.onmessage to handle the background container initially being empty. No point in checking each time the background changes if that can only happen the first time.