New Animation System!
Instead of calling update(), functions can request a certain kind of update job to done.

The addAnimationJob function then adds that job to a queue along with the animations type, start time, length, and end time. It then sets the animation state to active insure that subsequent calls to the function don't create new, unneeded loops.

If there's animation jobs in the queue then will loop through the jobs and check what types of job each is looking for and does it.

In order to insure that things aren't being drawn more than they need to for the active, each type of job just sets a redraw flag if it needs to be redrawn. If any of the jobs need to redraw, after all the jobs are looped through, it will then redraw the frame once.

Because of this, I was able to add a background fading animation.

Unrelated to all of that, I also made some slight adjustments to the resize function to make it do a little less work per-surface. I feel like there still might be savings to be had but I'm bad at math so don't get your hopes up lol