Set Canvas Sizes Again
Just realized that we only ever set the canvas size for the stage and we never took DPI scaling into consideration so everything else was rendering at the default 300x300 resolution. On 4K displays with 150% DPI scaling, all surfaces will be more expensive to render to except for the search bar which is cheaper. On a 1440p display without scaling, the dash is also cheaper than it was before.

Overall it's performance loss but that's fine because there's no good ways to benchmark a shader in WebGL2 so pushing it to render at native resolution and checking if it drops frames is the next best thing. It IS good to see that it looks so good at mismatched resolutions before though.