Send light list length with buffer
Before the light list and it's length would get sent separately but one must always be updated with the other so it makes sense to combine them into one buffer. 
The downside is that this requires more padding and the length needs to be sent as a float and cast to an int in the shader.

Hopefully there's a way around these issues, particularly the int() issue. I managed to at least make it look clean by reversing the loop. Hopefully that makes up for some of the performance loss.

Hopefully I can find some other use for the padding as there's currently a lot. There's three floats of padding per-light. When Z is added to each light, that at least cuts it down to two but that's still 25% of each light and having padding for the int brings it above 25% for the whole buffer.

Eventually the surface's Z will get taken into account so that least could one of the new slots of padding but it will only be worth it if we can update it's entry without updating the whole buffer. If that can happen then we can add surface brightness to the uniform buffer, too.