Convert Temp Canvas with CPU
When the a temporary canvas is created to analyze the image, we used to create an WebGL2 context and a texture to convert it to OkLab. Instead we're using a 2D context, writing the ImageBitmap to it, converting to image data, and running the de-gamma-ing and color space transform on the CPU. The reason for this is that it's less code and there aren't very many pixels anyway. Ideally we could convert the ImageBitmap to a Uint8Array directly.