Discovered IFFEs
...and thought to add functions to the window object.

Now imports aren't needed, there are fewer variables in the global namespace in js.js and each section has it's own private memory.
I also realized that appendChild returns the child so each section doesn't even stored in a constant.

It's just an overall re-organization of the code that makes it much harder for accidental mistakes to happen in the future. I'm not sure if anything is technically faster but the only things that were really changed a lot only happen once. Everything else should be just as quick. Since addAnimationJob() has it's own private memory now instead of using an object, it may even be slightly faster.