Now oneCanvas really does have one canvas
Now all surfaces are rendered to one canvas along with rounded corners. Top surfaces just look through to the canvas.

The search bar and stage don't need any special handling but the dropdown and menu and dash need look through to the canvas by feeding SVGs with the same dimensions into the image-mask property. Right now it's only being done on the dropdown but it needs some modification before applying to the dash. Performance is the best among all of the single canvas implementations. Changing themes and backgrounds seems to be at full frame rate but going between layouts isn't as smooth. I think it's because the layout is being re-flowed.